requirements.txt
streamlit
fastapi
uvicorn
langchain
langchain-community
langchain-openai
langchain-huggingface
pypdf
faiss-cpu
sentence-transformers
python-dotenv
python-multipart